fix: 순환참조 해결을 위해 서비스 클래스를 2개로 나누기
