+++ src/main/java/Team02/BackEnd/controller/CalendarController.java
@@ -5,7 +5,7 @@ import static Team02.BackEnd.constant.Constants.ACCESS_TOKEN_REPLACEMENT;
 
 import Team02.BackEnd.apiPayload.ApiResponse;
 import Team02.BackEnd.apiPayload.code.status.SuccessStatus;
-import Team02.BackEnd.service.CalendarService;
+import Team02.BackEnd.service.calendar.CalendarCheckService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -19,7 +19,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/spring")
 public class CalendarController {
 
-    private final CalendarService calendarService;
+    private final CalendarCheckService calendarCheckService;
 
     @GetMapping("/calendars")
     @Operation(summary = "달력 데이터", description = "쿼리 파라미터로 년,월 제공 => getDatesWhenUserId[답변 한 날짜] = answerId, 답변 안한 날짜는 0")
@@ -28,6 +28,6 @@ public class CalendarController {
                                                    @RequestParam("month") final String month) {
         String accessToken = authorizationHeader.replace(ACCESS_TOKEN_PREFIX, ACCESS_TOKEN_REPLACEMENT);
         return ApiResponse.of(SuccessStatus.GET_DATES_WHEN_USER_DID,
-                calendarService.getDatesWhenUserDid(accessToken, year, month));
+                calendarCheckService.getDatesWhenUserDid(accessToken, year, month));
     }
 }
