+++ src/main/java/Team02/BackEnd/controller/InsightController.java
@@ -5,7 +5,8 @@ import Team02.BackEnd.apiPayload.code.status.SuccessStatus;
 import Team02.BackEnd.converter.InsightConverter;
 import Team02.BackEnd.dto.insightDto.InsightRequestDto;
 import Team02.BackEnd.dto.insightDto.InsightResponseDto;
-import Team02.BackEnd.service.InsightService;
+import Team02.BackEnd.service.insight.InsightCheckService;
+import Team02.BackEnd.service.insight.InsightService;
 import io.swagger.v3.oas.annotations.Operation;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -24,6 +25,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class InsightController {
 
     private final InsightService insightService;
+    private final InsightCheckService insightCheckService;
 
     @PostMapping("/insights")
     @Operation(summary = "인사이트 저장하기 react -> spring", description = "질문요청에서 받은 answerId로 쿼리 파라미터")
@@ -36,7 +38,7 @@ public class InsightController {
     @GetMapping("/insights")
     @Operation(summary = "인사이트 받아오기 react -> spring", description = "질문요청에서 받은 answerId로 쿼리 파라미터")
     public ApiResponse<InsightResponseDto.GetInsightDto> getAiInsight(@RequestParam("answerId") final Long answerId) {
-        List<String> insight = insightService.getAiInsight(answerId);
+        List<String> insight = insightCheckService.getAiInsight(answerId);
         return ApiResponse.of(SuccessStatus.GET_INSIGHT, InsightConverter.toGetInsightDto(insight));
     }
 }
