+++ src/main/java/Team02/BackEnd/controller/QuestionController.java
@@ -8,8 +8,8 @@ import Team02.BackEnd.apiPayload.code.status.SuccessStatus;
 import Team02.BackEnd.converter.QuestionConverter;
 import Team02.BackEnd.domain.Question;
 import Team02.BackEnd.dto.questionDto.QuestionResponseDto;
-import Team02.BackEnd.service.AnswerService;
-import Team02.BackEnd.service.QuestionService;
+import Team02.BackEnd.service.answer.AnswerService;
+import Team02.BackEnd.service.question.QuestionCheckService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,7 +22,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/spring")
 public class QuestionController {
 
-    private final QuestionService questionService;
+    private final QuestionCheckService questionCheckService;
     private final AnswerService answerService;
 
     @GetMapping("/questions")
@@ -30,8 +30,8 @@ public class QuestionController {
     public ApiResponse<QuestionResponseDto.GetQuestionDto> getQuestion(
             @RequestHeader("Authorization") final String authorizationHeader) {
         String accessToken = authorizationHeader.replace(ACCESS_TOKEN_PREFIX, ACCESS_TOKEN_REPLACEMENT);
-        Question question = questionService.getUserQuestion(accessToken);
-        Long answerId = answerService.getAnswerId(accessToken, question);
+        Question question = questionCheckService.getUserQuestion(accessToken);
+        Long answerId = answerService.createAnswer(accessToken, question);
         return ApiResponse.of(SuccessStatus.GET_QUESTION, QuestionConverter.toQuestionResponseDto(question, answerId));
     }
 }
