+++ src/main/java/Team02/BackEnd/controller/RecordController.java
@@ -6,8 +6,8 @@ import static Team02.BackEnd.constant.Constants.ACCESS_TOKEN_REPLACEMENT;
 import Team02.BackEnd.apiPayload.ApiResponse;
 import Team02.BackEnd.apiPayload.code.status.SuccessStatus;
 import Team02.BackEnd.dto.recordDto.RecordRequestDto;
-import Team02.BackEnd.service.FeedbackService;
-import Team02.BackEnd.service.UserService;
+import Team02.BackEnd.service.feedback.FeedbackService;
+import Team02.BackEnd.service.user.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
