+++ src/main/java/Team02/BackEnd/service/insight/InsightService.java
@@ -1,8 +1,9 @@
-package Team02.BackEnd.service;
+package Team02.BackEnd.service.insight;
 
 import Team02.BackEnd.domain.Answer;
 import Team02.BackEnd.domain.Insight;
 import Team02.BackEnd.repository.InsightRepository;
+import Team02.BackEnd.service.answer.AnswerCheckService;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -13,11 +14,11 @@ import org.springframework.stereotype.Service;
 @Slf4j
 public class InsightService {
 
+    private final AnswerCheckService answerCheckService;
     private final InsightRepository insightRepository;
-    private final AnswerService answerService;
 
     public void saveAiInsight(final List<String> insights, final Long answerId) {
-        Answer answer = answerService.getAnswerByAnswerId(answerId);
+        Answer answer = answerCheckService.getAnswerByAnswerId(answerId);
         insights.stream()
                 .map(insight -> Insight.builder()
                         .insight(insight)
@@ -25,13 +26,5 @@ public class InsightService {
                         .build())
                 .forEach(insightRepository::save);
         log.info("질문에 대한 AI 인사이트 생성, answerId : {}", answerId);
-    }
-
-    public List<String> getAiInsight(final Long answerId) {
-        log.info("질문에 대한 AI 인사이트 가져오기, answerId : {}", answerId);
-        Answer answer = answerService.getAnswerByAnswerId(answerId);
-        return insightRepository.findAllByAnswerId(answer.getId()).stream()
-                .map(Insight::getInsight)
-                .toList();
     }
 }
