+++ src/main/java/Team02/BackEnd/service/question/QuestionCheckService.java
@@ -1,10 +1,15 @@
-package Team02.BackEnd.service;
+package Team02.BackEnd.service.question;
 
 import Team02.BackEnd.apiPayload.code.status.ErrorStatus;
 import Team02.BackEnd.apiPayload.exception.handler.QuestionHandler;
+import Team02.BackEnd.domain.Answer;
 import Team02.BackEnd.domain.Question;
 import Team02.BackEnd.domain.oauth.User;
 import Team02.BackEnd.repository.QuestionRepository;
+import Team02.BackEnd.service.answer.AnswerCheckService;
+import Team02.BackEnd.service.feedback.FeedbackCheckService;
+import Team02.BackEnd.service.user.UserCheckService;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -12,15 +17,17 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class QuestionService {
+public class QuestionCheckService {
 
+    private final UserCheckService userCheckService;
+    private final AnswerCheckService answerCheckService;
+    private final FeedbackCheckService feedbackCheckService;
     private final QuestionRepository questionRepository;
-    private final UserService userService;
-    private final AnswerService answerService;
 
     public Question getUserQuestion(final String accessToken) {
-        User user = userService.getUserByToken(accessToken);
-        if (answerService.isWrongAnswer(user)) {
+        User user = userCheckService.getUserByToken(accessToken);
+        Optional<Answer> latestAnswer = answerCheckService.getLatestAnswerByUser(user);
+        if (latestAnswer.isPresent() && feedbackCheckService.isFeedbackExistsWithAnswer(latestAnswer.get())) {
             user.minusQuestionNumber();
         }
         Question question = getQuestionByUserQNumber(user.getQuestionNumber());
