+++ src/main/java/Team02/BackEnd/service/selffeedback/SelfFeedbackCheckService.java
@@ -1,13 +1,13 @@
-package Team02.BackEnd.service;
+package Team02.BackEnd.service.selffeedback;
 
 import Team02.BackEnd.apiPayload.code.status.ErrorStatus;
 import Team02.BackEnd.apiPayload.exception.handler.SelfFeedbackHandler;
-import Team02.BackEnd.converter.SelfFeedbackConverter;
 import Team02.BackEnd.domain.Answer;
 import Team02.BackEnd.domain.SelfFeedback;
 import Team02.BackEnd.domain.oauth.User;
-import Team02.BackEnd.dto.selfFeedbackDto.SelfFeedbackRequestDto.SaveSelfFeedbackDto;
 import Team02.BackEnd.repository.SelfFeedbackRepository;
+import Team02.BackEnd.service.answer.AnswerCheckService;
+import Team02.BackEnd.service.user.UserCheckService;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Objects;
@@ -15,31 +15,18 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-@Service
 @RequiredArgsConstructor
 @Slf4j
-public class SelfFeedbackService {
+@Service
+public class SelfFeedbackCheckService {
 
+    private final UserCheckService userCheckService;
+    private final AnswerCheckService answerCheckService;
     private final SelfFeedbackRepository selfFeedbackRepository;
-    private final UserService userService;
-    private final AnswerService answerService;
-
-    public void saveSelfFeedback(final Long answerId, final SaveSelfFeedbackDto saveSelfFeedbackDto) {
-        Answer answer = answerService.getAnswerByAnswerId(answerId);
-        if (isExistsSelfFeedback(answerId)) {
-            SelfFeedback selfFeedback = getSelfFeedbackByAnswerId(answerId);
-            selfFeedback.updateFeedback(saveSelfFeedbackDto.getFeedback());
-            selfFeedbackRepository.save(selfFeedback);
-            return;
-        }
-        SelfFeedback selfFeedback = SelfFeedbackConverter.toSelfFeedback(answer, saveSelfFeedbackDto);
-        selfFeedbackRepository.save(selfFeedback);
-        log.info("스피치에 대한 셀프 피드백 저장, selfFeedbackId : {}", selfFeedback.getId());
-    }
 
     public SelfFeedback getLatestSelfFeedback(final String accessToken) {
-        User user = userService.getUserByToken(accessToken);
-        List<Answer> answers = answerService.getAnswersByUser(user);
+        User user = userCheckService.getUserByToken(accessToken);
+        List<Answer> answers = answerCheckService.getAnswersByUser(user);
         SelfFeedback selfFeedback = answers.stream()
                 .map(answer -> selfFeedbackRepository.findByAnswerId(answer.getId()))
                 .filter(Objects::nonNull)
@@ -50,11 +37,11 @@ public class SelfFeedbackService {
         return selfFeedback;
     }
 
-    private SelfFeedback getSelfFeedbackByAnswerId(final Long answerId) {
+    public SelfFeedback getSelfFeedbackByAnswerId(final Long answerId) {
         return selfFeedbackRepository.findByAnswerId(answerId);
     }
 
-    private boolean isExistsSelfFeedback(final Long answerId) {
+    public boolean isExistsSelfFeedback(final Long answerId) {
         return selfFeedbackRepository.findByAnswerId(answerId) != null;
     }
 
@@ -64,4 +51,3 @@ public class SelfFeedbackService {
         }
     }
 }
-
