+++ src/test/java/Team02/BackEnd/controller/ReissueControllerTest.java
@@ -5,7 +5,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 
 import Team02.BackEnd.jwt.service.JwtService;
-import Team02.BackEnd.service.ReissueService;
+import Team02.BackEnd.service.reissue.ReissueService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
